+++ small_text/query_strategies/__init__.py
@@ -6,12 +6,15 @@ from small_text.query_strategies.coresets import (greedy_coreset,
                                                   GreedyCoreset,
                                                   lightweight_coreset,
                                                   LightweightCoreset)
-from small_text.query_strategies.strategies import (QueryStrategy,
-                                                    RandomSampling,
-                                                    ConfidenceBasedQueryStrategy,
-                                                    BreakingTies,
-                                                    LeastConfidence,
-                                                    PredictionEntropy,
-                                                    SubsamplingQueryStrategy,
-                                                    EmbeddingBasedQueryStrategy,
-                                                    EmbeddingKMeans)
+from small_text.query_strategies.strategies import (
+    QueryStrategy,
+    RandomSampling,
+    ConfidenceBasedQueryStrategy,
+    BreakingTies,
+    LeastConfidence,
+    PredictionEntropy,
+    SubsamplingQueryStrategy,
+    EmbeddingBasedQueryStrategy,
+    EmbeddingKMeans,
+    ContrastiveActiveLearning
+)